Add files via upload
The aim of this R Markdown is to comprehensively show each step of the k-cluster analysis used for the survey of communicaTUM. Before pursuing the final k-means algorithm, we need to choose how many clusters (k) we want. We want the number of clusters to represent the German population the best, while still being plausible. Therefore, we need to do an optimization process in order to see, what cluster number k gives the best results. Lets go to the code